[WIP]change dir
