+++ config/mac.spec
