+++ config/win.spec
